refactor(card): move everything to Card & remove BulletList
affects: @fremtind/jkl-card-react
