--- packages/card-react/src/CardAction.tsx
@@ -1,9 +1,0 @@
-import { MouseEventHandler } from "react";
-
-export type validButtons = "primary" | "secondary" | "tertiary";
-
-export type Action = {
-    type: validButtons;
-    name: string;
-    onClick: MouseEventHandler<HTMLButtonElement>;
-};
